refactor: remove legacy init and scaffold commands while adding new CreateUser usecase files
