+++ src/app/usecase/mod.rs
@@ -6,9 +6,9 @@ pub mod validate_project;
 pub use generate_artifacts::{GenerateArtifactsInput, GenerateArtifactsUseCase};
 pub use init_project::{InitProjectInput, InitProjectUseCase};
 pub use plan_architecture::{
-	PlanArchitectureInput,
-	PlanArchitectureOutput,
-	PlanArchitectureUseCase,
-	PlanRenderFormat,
+    PlanArchitectureInput,
+    PlanArchitectureOutput,
+    PlanArchitectureUseCase,
+    PlanRenderFormat,
 };
 pub use validate_project::{ValidateProjectInput, ValidateProjectUseCase};
