+++ src/application/usecases/create_user.prompt.md
@@ -1,0 +1,1 @@
+<!-- Prompt override for: create_user (usecase) -->
