+++ src/application/usecases/create_user.rs
@@ -1,0 +1,1 @@
+// Batonel placeholder: create_user (usecase)
