+++ src/commands/mod.rs
@@ -4,12 +4,10 @@ pub mod fix;
 pub mod fix_rollout;
 pub mod triage;
 pub mod guard;
-pub mod init;
 pub mod plan;
 pub mod policy_resolve;
 pub mod preset_migrate;
 pub mod preset_registry;
 pub mod preset_verify;
 pub mod prompt;
-pub mod scaffold;
 pub mod verify;
